Removed unnecessary code after implementation of the new authentication design.
